timetrack_pro/requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
Faker>=22.0,<23.0

# Linting & Formatting
ruff>=0.2,<1.0

# Type Checking
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Security
safety>=2.3,<3.0

# Development utilities
ipython>=8.20,<9.0
ipdb>=0.13,<1.0
